Create Fully-Functional and Good-looking UI
Google Generative AI Chat Function is used, to do a Chat, rather than it generateContent Function.

The Server API is given a new user Prompt as well as old Chat History. Both of which are passed to Gemini API.

The UI Design is created for the Chat. The Design is Clean.
More Great things are added:
 - The Textarea is multi-Line and changes its height alongwith Content. If reached certain height, scroll-bar appears.
 - The Chat Bars are of different Colors, and has top-side corner-border, rather than round, to make a look of Arrow.

 - User Messages are floated towards Right.
 - The Message Section gets a Scroll-bar, if its content exceeds.

 - Custom Scrollbars are styled in globals.css.
 - Padding for Small Screen are also set cleanly.

Initially, the chat History contains some Chats. Gemini dont accept the chatHistory to have model's Text first.